Move utils in a single module.
