+++ linguist/models/__init__.py
@@ -1,9 +1,9 @@
 # -*- coding: utf-8 -*-
 from .. import settings
-from ..utils.models import load_class
+from .. import utils
 
 
-Translation = load_class(settings.TRANSLATION_MODEL)
+Translation = utils.load_class(settings.TRANSLATION_MODEL)
 
 
 from ..signals import *  # noqa
